+++ docs/ui/ShinyText.tsx
